backend/requirements.txt
fastapi
uvicorn
aiohttp
aiosqlite
aioredis
python-whois
ipwhois
python-dotenv
